+++ src/main/java/com/example/libraryviewerbackend/exceptions/ObjectNotFoundException.java
@@ -4,9 +4,9 @@ import lombok.Getter;
 
 @Getter
 public class ObjectNotFoundException extends RuntimeException{
-    private final Integer objectId;
+    private final Long objectId;
     private final String objectType;
-    public ObjectNotFoundException(String message, Integer objectId, Class<?> objectClass) {
+    public ObjectNotFoundException(String message, Long objectId, Class<?> objectClass) {
         super(message);
         this.objectId = objectId;
         this.objectType = objectClass.getSimpleName();
